ItemDoPedidoDAO e população do banco de dados

+++ src/projetofinaljava/ItemDoPedido.java
@@ -10,19 +10,19 @@ package projetofinaljava;
  * @author sharonhasegawa
  */
 public class ItemDoPedido {
-    private int quantidade;
+    private long quantidade;
     private Produto produto;
 
-    public ItemDoPedido(int quantidade, Produto produto) {
+    public ItemDoPedido(long quantidade, Produto produto) {
         this.quantidade = quantidade;
         this.produto = produto;
     }
 
-    public int getQuantidade() {
+    public long getQuantidade() {
         return quantidade;
     }
 
-    public void setQuantidade(int quantidade) {
+    public void setQuantidade(long quantidade) {
         this.quantidade = quantidade;
     }
